requirements.txt
aioari==0.10.2
aioredis==1.3.1
aioswagger11==0.9.1
amqp==5.1.1
appdirs==1.4.4
arabic-reshaper==2.1.3
argon2-cffi==21.1.
Django==3.1.7
django-compression-middleware==0.5.0
django-cors-headers==3.7.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-database-size==1.0.2
django-ipware==3.0.2
django-jsonfield==1.4.1
django-logging-json==1.15
django-modeltranslation==0.17.1
django-parler==2.2
django-parler-rest==2.1
django-requestlogs==0.2.3
django-rest-framework==0.1.0
django-rest-multiple-models==2.1.3
django-rest-swagger==2.2.0
django-searchable-encrypted-fields==0.1.9
django-simple-history==3.0.0
djangorestframework==3.12.2
docopt==0.6.2
drf-yasg==1.20.0
